pyproject.toml
[tool.poetry]
name = "apxivibetts-bot"
version = "0.1.0"
description = "Telegram бот для озвучки тексту через Gemini TTS"
authors = ["ApXiVibe <you@example.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
python-telegram-bot = ">=20,<22"
python-dotenv = "^1.0.1"
pydub = "^0.25.1"
google-genai = "^0.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
ruff = "^0.6.8"
black = "^24.10.0"
mypy = "^1.11.0"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
lint.select = ["E", "F", "I", "N", "PL", "UP", "S", "B"]
lint.ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "PLR2004", "F401"]
"bot.py" = ["PLR0912"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_any_generics = true

